+++ src/components/generic/editable-field-addables/index.tsx
@@ -1,4 +1,4 @@
-export { default as MetaformTextFieldComponent } from "./metaform-text-field-component";
+export { default as MetaformTextFieldAddable } from "./text-field-addable";
 export { default as MetaformHtmlAddable } from "./html-field-addables";
 export { default as MetaformRadioFieldAddable } from "./radio-field-addable";
 export { default as MetaformSubmitFieldComponent } from "./metaform-submit-field-component";
